Meson: Don't create libsysdeps_xxx.a, just add the file listing
Simplifies the build process and incidentally also fixes the
inconsistency in build flags -- the files from sysdeps/ weren't getting
the march_flags, for example.

Those `-DLINUX` and `-D__linux__` flags were there since the Meson build
system was added to the internal repository prior to open-sourcing, but
they don't seem to have a purpose because a) nothing uses the `LINUX`
macro and b) the compiler predefines the `__linux__` macro for us.

Drive-by addition of the copyright and SPDX headers to the Darwin and
FreeBSD files.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>