FW/SharedMemory: stop passing never-needed state to the child
`force_test_time` and `test_time` were never used in the child that I
can see, even in the pre-Open Source version. It's used in the
`test_duration()` function, but that function is only in the parent.

Likewise for `max_test_time` in `test_timeout()` and timeouts are
definitely a parent-side thing.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>