FW/Logging: propagate the full results from logging_print_results()
Up until this point, it only returned Skipped, Passed, or Failed. Now it
will return the full status, with the near-term objective of making it
summarise the result from multiple child processes.

This in turn means run_one_test_once() can return extended results to
run_one_test(). The latter will not, because at the end of the failure
analysis it sets to a single condition:

```c++
    /* now we process retries */
    if (fail_count > 0) {
...
        analyze_test_failures(*tc, test, fail_count, iterations, per_cpu_fails);
        state = TestResult::Failed;
    }
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>