FW/Logging: refactor TapFormatLogger::print() to better use C++
This code was originally written in C, which is why it manipulates two
raw pointers. The last version of the logging_print_results() function
in C (dated 2020-07-30) still received that `const char *extra` as a
parameter.

So it's high time to clean it up and remove some of the fragility. For
example, if the `ChildExitStatus::extra` value isn't 0, this code was
assuming the result was a crash of some sort, which made me go down the
wrong path of debugging to find out why the selftest_pass test was
crashing... it wasn't, the log was wrong.

Last C version:
```c
bool logging_print_results(int pc, int tc, struct test *test, const char *extra)
{
    bool failed = (pc >= 0 && (unsigned)pc != (unsigned)thread_count);
    const char *qual = quality_string(test);
    if (qual[0] == '\0' && extra[0] != '\0')
        qual = "# ";

    switch (output_format) {
    case key_value_format:
        logging_printf(LOG_LEVEL_QUIET, "%s_result = %s\n", test->id,
                       pc < 0 ? "skip" : failed ? "fail" : "pass");
        logging_printf(LOG_LEVEL_VERBOSE(1), "%s_seq = %d\n", test->id, tc);
        logging_printf(LOG_LEVEL_VERBOSE(1), "%s_quality = %s\n", test->id, qual);
        logging_printf(LOG_LEVEL_VERBOSE(1), "%s_description = %s\n", test->id, test->name);
        logging_printf(LOG_LEVEL_VERBOSE(1), "%s_pass_count = %d\n", test->id, pc);
        if (pc >= 0)
            logging_printf(LOG_LEVEL_VERBOSE(1), "%s_fail_percent = %.1f\n", test->id,
                           100. * (thread_count - pc) / thread_count);
        break;

    case tap_format:
        if (failed)
            logging_printf(LOG_LEVEL_QUIET, "not ok %2i %-16s %s %s\n", tc, test->id, qual, extra);
        else
            logging_printf(LOG_LEVEL_VERBOSE(1), "ok %2i %-20s %s\n", tc, test->id, extra);
        break;
    }

    logging_flush();
    print_thread_messages(test);
    logging_flush();
    return !failed;
}
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>