FW/SharedMemory: move init_shmem() to further up in main()
I'm about to move the contents of SandstoneApplication::ExecState to the
shared memory segment, so said segment must exist by the time we're
parsing the command-line. To simplify the execution, I opted to always
create a temporary file, which we simply close later if unneeded.

This means Linux systems without memfd (introduced in 3.17, so read "Red
Hat Enterprise Linux 7", unless they backported the feature) will create
an extra file, but since this is not the last one anyway, it hardly
matters. It may even be better because we find out very early if the
creation of temporary files fails.

This does not affect Windows.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>