CI/Bats: Build the Windows binary on Windows
The upgrade to Fedora 41 was not a good idea. The WINE in that system is
giving us trouble and I haven't been able to figure out what it
was. Instead, let's run on Windows for real.

To do that, we need to switch to UCRT, because the MSVCRT-based build
was having problems, like for example the `tmpfile()` function
failing. Since Fedora doesn't have UCRT build packages, we have to build
*on* Windows now.

I had resisted doing that because it used to take much longer to build
on Windows, especially package installation, but it seems to have
improved.

| Step                    | Cross-compiling | Native |
| Installing dependencies | 32 + 44 =  76 s |  182 s |
| Compilation             | 2 + 171 = 173 s |  285 s |
| Run tests               |           187 s |  100 s |
| Total                   |           436 s |  567 s |

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>