FW: make LogicalProcessorSet dynamically-sized
This allows us to go beyond the 1024 logical processor limit that the
tool currently has. That number is very close to what an 8-socket
Sapphire Rapids may have (896 for 56 cores per socket, 960 for 60 cores
per socket) and is exceeded for servers with more than 8 sockets.

One more MAX_THREADS user bites the dust!

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>