FW/Logging: keep the file descriptors open across tests
Instead of closing and recreating more memfds, we truncate them back to
zero and seek to the beginning. Unproven, but it feels like this should
be faster: it's the same number of system calls between tests, but
truncating and seeking can't be as expensive as a single new file-
opening.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>