FW: add proper test slice planning
This concludes the main development of the "socket separation" feature,
by finally implementing parallel-slicing by default for all tests. There
are still some clean-ups and minor tweaks that may go in.

For example, the formula as implemented in this commit probably deserves
some attention. For a CPU like the 144-core Sierra Forest, the ideal
plan would be 6 slices of 24 cores, but the formula calculates 5 slices
(4x32 and one of 16).

The logging output may still change too.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>