FW/Linux: don't keep trying to open the /dev/ MSR nodes if we get EACCES
We can't gain privileges, so the result won't change if we keep trying.

Not retrying means we avoid system calls. It's not a lot, but at 2xNPROC
system calls on start, even 0.5µs accummulates. This patch saves 2 ms in
start up time with no tests run and 20 ms before starting the first test.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>