FW: move the between-test checks on MCE and thermal events to mce_check
We were only producing comments in the output, which means many tools
never caught stored them at all. So, instead, just move to the mce_check
test. As a consequence the test running in a child process, its update
of the counts won't be committed, so when the test is run again, it'll
compare against the original numbers from the tool's start.

[ChangeLog][mce_check] This test now has a "sticky" behavior: if a
problem is found, it is reported every time the test is run, instead of
only the first time the problem was detected.

This change removes the inspecting of /proc/interrupts from between each
test, which turned out to be a major time contributor, as it ran
completely serially between tests. A benchmark of running the
selftest_pass test 1000 times reveals the improvement:

| Machine     | Before (ms) | After (ms) |
|-------------|-------------|------------|
| 14c SKX     |       6.83  |      6.52  |
| 2x28c CLX   |      21.45  |     13.91  |
| 2x48c SPR   |      63.73  |     17.23  |

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>