FW: pre-allocate the stacks' threads
There are two reasons for this. First, it will allow us to randomise the
stack addresses according to the fracture, replacing the code that was in
`run_one_test()` and didn't do anything. Second, because it shows a
performance improvement:

|  # Threads   | Before (µs)  |  After (µs)  |
|--------------|--------------|--------------|
|            1 |        35950 |        33363 |
|           24 |        53527 |        53386 |
|           48 |        57227 |        55624 |
|           96 |        62519 |        59104 |
|          192 |        74021 |        67249 |

Subtracting the first line from all others:

|            |        Before (µs)      |        After (µs)       |
|  # Threads |      Total | Per thread |      Total | Per thread |
|------------|------------|------------|------------|------------|
|         23 |      17576 |        764 |      20022 |        871 |
|         47 |      21277 |        453 |      22261 |        474 |
|         95 |      26568 |        280 |      25741 |        271 |
|        191 |      38071 |        199 |      33886 |        177 |

Even subtracting the overhead of starting a thread, the marginal cost of
starting threads continues to decrease (I'm not sure why, I expected it
to increase).

I also benchmarked not having guard pages, but that seems to trigger a
perverse optimisation inside the kernel, at least in the machine I
tested: it decided to offer 2 MB huge pages, but unfortunately that's
slower to allocate.

Pre-faulting the top of the stack for each doesn't seem to help, because
`pthread_create()` writes there anyway.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>