FW: remove the useless code allocating random memory in run_one_test()
This code is running in the main thread and allocates less than a page,
so it is served from the main thread's heap. That means this doesn't
affect the stacks or heaps of any of the test threads, nor will it
affect memory allocated by the tests' init() functions as that is also
run in a separate thread since 4248649b550477eb299efcdbfd35be85bbb11657.

See #333 for a better solution for stacks.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>