FW/Topology: resize Topology::packages only once
Since 9ed9a0c6c1c87304f80af746e2e996803f86304d (PR #321), the `cpu_info`
array is sorted, so we know the highest numbered package is last.
Additionally, because we sort in unsigned, we know that any unknown
package IDs will be stored last too with an id of -1.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>