FW/Win32: Provide C++ operators new and delete
The ones in libstdc++ do call malloc and aligned_alloc and other
suitable functions we do provide. However, on Windows, that only works
if you link statically to libstdc++.a. If you're using libstdc++6.dll,
the already-linked DLL will instead go to the runtime (either MSVCRT or
UCRT), bypassing our allocating and zeroing functions.

All the functions in this file are the same because we never throw and
never return null pointers.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>