FW/Logging: move the increment of messages_logged into a common place
Into `log_message_for_thread()`, which is called by
`log_message_preformatted()`, `log_data_common()`,
`logging_format_data()` and `logging_format_data()`. The former two
incremented the count, but the latter two didn't.

This means we aren't atomically deciding whether to log or not, but
that's acceptable because cross-thread logging is not expected to be
thread anyway.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>