FW/Skip reasons: Add DummySkipCategory to skip reasons
This category can be used when majority of the cores skip a test.
When this skip category is used, Sandstone will not log any skip messages.
This is equivalent to the case when all cores run the test and the test
exits without skips.

One of the use cases for this are the accelerator tests where only a
handful of cores run the accelerator tests and others skip. By default
all cores will log their defualt/user defined skip messages, the final
log looks messy. With the changes in this commit, when this skip category
is used, Sandstone will not log the default/user defined skip log message.

Before these changes.

test: simple_add
  time-at-start: { elapsed:   1183.377, now: !!timestamp '2023-08-01T04:33:24Z' }
  result: pass
  time-at-end:   { elapsed:   1193.377, now: !!timestamp '2023-08-01T04:33:24Z' }
  test-runtime: 9.056
  threads:
  - thread: 0
    id: { logical:  0, package: 0, core:  0, thread: 0, family: 6, model: 0x55, stepping: 6, microcode: 0x400320a, ppin: null }
    messages:
    - { level: skip, text: 'Os not supported category' }
  - thread: 2
    id: { logical:  2, package: 0, core:  2, thread: 0, family: 6, model: 0x55, stepping: 6, microcode: 0x400320a, ppin: null }
    messages:
    - { level: skip, text: 'Os not supported category' }

After these changes.

- test: simple_add
  time-at-start: { elapsed:    993.370, now: !!timestamp '2023-08-01T00:05:02Z' }
  result: pass
  time-at-end:   { elapsed:   1010.037, now: !!timestamp '2023-08-01T00:05:02Z' }
  test-runtime: 16.848
- test: simple_add
  time-at-start: { elapsed:   1023.371, now: !!timestamp '2023-08-01T00:05:02Z' }
  result: pass
  time-at-end:   { elapsed:   1043.371, now: !!timestamp '2023-08-01T00:05:02Z' }
  test-runtime: 19.000
- test: simple_add
  time-at-start: { elapsed:   1053.372, now: !!timestamp '2023-08-01T00:05:02Z' }
  result: pass
  time-at-end:   { elapsed:   1073.373, now: !!timestamp '2023-08-01T00:05:02Z' }
  test-runtime: 18.406

Signed-off-by: Dilip Shivaraju <dilip.shivaraju@intel.com>