ifs: add a preinit function to set the minimum duration
Because this test is sequential, its runtime is proportional to the
number of cores being tested. The time for each Scan test is at most 200
ms (per Intel documentation). This will cause the test's target time to
be set at a minimum to this value, which then will inform the test's
timeout.

We're not bothering to try and find the exact number of cores that are
going to be tested, so we may overestimate the minimum duration by half.

This will make a difference for a system with more than 225 logical
processors because the formula in `test_timeout()` will go above the
5-minute default:

```c++
    ShortDuration result = regular_duration * 5 + 30s;
    if (result < 300s)
        result = 300s;
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>