CI: add a test for skipping with unsatisfied .minimum_cpu
In the current x86-64 build, we are using all bits of the uint64_t, so
it is possible there will be a CPU with all those bits set in the
future, but highly unlikely there will be one with `cpu_feature_hybrid`
and either `cpu_feature_avx512f or `cpu_feature_pconfig`.

I'm also seeing the need to extend this to 128 bits soon.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>