CI/Bats: correct the regexp to match socket1 failures
The regexp was written when the mock topology was "0 1 2 3" (like in the
"triage" test), that is, one core per socket. At some point before
merging, I changed that to a mock topology of 2Sx2c but failed to update
the regular expression to match.

This expression forces that the mask for all sockets except the second
be composed of dots only. For the second, it allows both X and ., but
requires at least one X (that also means it won't catch heterogeneous
results from hyperthreads).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>