FW/child_debug: ensure we drain the queue of crash information
This fixes a race condition, because the draining of the socket depended
on the order in which we'd get the SIGCHLD / pidfd activation versus the
socket. I hadn't observed this on Linux, but it can happen there two: if
the parent process is woken up after two children have crashed, we'd
only get the context of one.

If this race condition happened, we'd probably blame the next process as
a crash too. While that doesn't detection worse (we've already had a
crash), it could complicate debugging of the part.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>