FW/child_debug: add "info registers" gdb command outside of x86-64
For x86-64, we have the context dumper in `dump_gprs()`, which we use
preferentially because it doesn't require starting the debugger (which
takes time parsing the debug symbols, if any). For other platforms,
since we don't have that, let's ask gdb to print the register state for
us.

Implementing said functionality for other platforms would be a welcome
contribution to OpenDCDiag.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>