FW/Win32: statelessly map LogicalProcessor to PROCESSOR_NUMBER
Instead of statefully depending on the previous count of members in each
group, just store them in the bits of the `LogicalProcessor` variable.
This turns the operations that used to be O(n) to O(1).

More importantly, this means we don't depend on stateful data that
needed to be initialised. It was only initialised by
`ambient_logical_processor_set()`, which the child process stopped
calling (probably in commit bbf1ad64067897803fdd612e992aa7c827bf1baa).
That was an intentional change, because the child should never need to
know the full system's ambient processor set, only what the parent
process tells it to care about.

As a side-effect of that, the child process wouldn't have the state
initialised, so `processor_group_for()` would always return zero and
`processor_in_group()` would return `n` (cast to `BYTE`). That is only
correct for systems with a single processor group, which happens to be
all the development machines we tested this on. For systems with more
than one group, we would fail to map back from `LogicalProcessor` to
`PROCESSOR_NUMBER`. Additionally, we'd cause Undefined Behaviour on
line:
```c++
     groupAffinity.Mask = KAFFINITY(1) << processorNumber.Number;
```
for logical processors 64 to 255, 320 to 511, etc.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>