FW/Win32: implement pin_to_logical_processors()
We're expecting that the child processes are restricted to a single
Windows Processor Group, in which case we can ask the OS to set the
affinity to that group. This will affect the child process' main thread
only, so it's not a big deal if we don't do anything. It's actually more
important that we be quick.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>