FW/Win32: fix kludge and increase maximum processor count
This fixes two problems in the Win32 `wait_for_children()`'s
`single_wait` lambda. First, the kludge introduced by commit
3ba26181c41a2aad9dd9fd97c8763fdf9ffb0933 that used a dummy handle to
make the multi-child array passed to WaitForMultipleObjects work when
one child had already exited. The second, the fact that
WaitForMultipleObjects is limited to MAXIMUM_WAIT_OBJECTS (64) objects
at a time (which is why we were limited to 2048 cores at maximum).

The documentation[1] for WaitForMultipleObjects suggests using threads
to wait on multiple WaitForMultipleObjects, which is overkill for us. We
can wait on a sliding window of first N children still pending to exit.
Even if one or a few of them time out, we should keep adding children to
the wait list and processing their exits. The only problem is if *all*
the N children are timing out but the next few aren't; in that case,
we'd blame all the still-unhandled children and say they timed out.
That's a very unlikely scenario.

[1] https://learn.microsoft.com/en-us/windows/win32/api/synchapi/nf-synchapi-waitformultipleobjects

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>