FW/Logging: add logging_flush() to logging_init()
This is right after we print the test that is about to start, so is a
logical place to flush the log to filesystem.

I noticed this on a defective machine that rebooted because of a test,
that the actual log file said a test had finished and passed:
```yaml
  time-at-start: { elapsed: 13592188.000, now: !!timestamp '2023-10-25T02:06:39Z' }
  result: pass
  time-at-end:   { elapsed: 13592280.000, now: !!timestamp '2023-10-25T02:06:39Z' }
  test-runtime: 92.027
```

But the terminal from which I launched the tool (tmux rules!) logged the
next test starting:
```yaml
- test: xxxxx
  time-at-start: { elapsed: 13592280.000, now: !!timestamp '2023-10-25T02:06:39Z' }
```

[ChangeLog][Framework] Fixed a shortcoming in flushing the logs that
could cause the start of a test to be missing if the system under test
crashed very soon after the test (only the end of the previous test
would be logged).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>