FW: Remove ResourceIssueSkipCategory.
This category was removed and all usages replaced with a more
appropriate ones. The enum value of 1 is being used by another category
now. Also, the order in the definition was re-shuffled, but since this
existed very briefly, we do not have cases relying on numerical value of
the enum (which should not be done anyways).

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>