FW: add cleanup_global() to centralize application exiting
We exited in three different places but only one duplicated the call to
`logging_close_global()`. Only the main exit remembered to call
`print_application_footer()` and restore any system state we modified
(in the current version, the `--vary-frequency` state).

This centralizes everything in a function that Does The Right Thing™.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>