FW/ContextDump: refactor to internally use std::string
I had used `FILE *` originally so this code could be useful in C
applications too (that's why it didn't use any of our own utilities
either). But let's use std::string instead because that will allow me to
get rid of `open_memstream()` and temporary files in the caller code.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>