FW/child_debug: use std::string for gdb communication too
We were using a plain file descriptor because that allowed us to use
`splice()` straight from the pipe to the backing file, in case the
backtrace was big. But that's really unnecessary and was a source of
bugs because the non-Linux code was never tested.

Replace with direct std::string and move up the location where we call
the logging functions to outside of the GDB communication loop.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>