FW/logging: don't print the "messages:" unnecessarily
The YAML mode output was printing it even if we weren't going to print
anything. This came about with the skip-reason functionality because the
skip reason is sent to the main thread log file.

NOT ENOUGH: Further refactoring to just put everything in large std::strings and
avoid system calls to be done later.

The TAP and key-value pair loggers are probably also affected, but I
don't care too much about that.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>