FW: Add ignore-test-failure mode.
This PR adds an experimental mode where the framework is expected to ignore errors raised by the tests.
Although no error is reported if any test fails, the failures are counted in `ignored_fails` counter for each thread.
When the test run is concluded a log with information can be printed to the console to infrom which threads have just failed and how many times.

Log example:

```
ok   n executed_test
 ---
  info: {version: <version>, timestamp: <timestamp>}
  Thread N on CPU M (pkg i, core j, thr k, family/model/stepping xx-xx-xx, microcode 0xnnnnnnn, PPIN N/A):
   - 'I> Found ignored failures. Number of ignored failures nnnn'
```

To run tests in mode where failures are ignored the framework needs to be compiled with Meson option: `ignore_test_fails=true`
Next run OpenDCDiag with parameter `ignore-test-failure`.

`ignore-test-failure` flag changes behavior of functions used to report test failures:
* report_failure
* report_failure_msg
* memcmp_or_fail
The functions are expected to do `return` to ensure test continuation.

Signed-off-by: dawid-sabat <dawid.sabat@intel.com>