FW: Refactor test management.
This replaces the entire test_selector infrastructure to pick the next
test. It encapsulates the function of iterating through tests into
SandstoneTestSet class whose only contract is to respond to the
question "what is the next test to run?". For this purpose, it provides
a configurable iterator. The following aspects can be configured:

    * Whether the tests are picked in random order or not;
    * Whether to "cycle though" the configured test set: an infinite
      iterator over the tests;
    * Whether to ignore test names that are not known (i.e. the test
      with that name cannot be found).

The special handling of the mce_check is removed from sandstone.cpp
(including injection on time period, before running next iteration).
Instead, mce_check is added to SandstoneTestSet and is being kept the
last on the list (e.g. in case of randomized access).

Additionally,

    * Current test count is kept globally and is not reset when the next
      iteration starts;
    * All special handling of mce_check test is removed from the main
      cycle and is encapsulated in the iterator instead.
    * The options to use built-in test list, test list file, as well as
      command-line options are now can be used together and combined.
      For example, an undesired test from a test list file can be
      disabled on the command-line using --disable option as opposed to
      editing the test list file needed before.
    * smi_count test and associated accounting is refactored (even
      though with the current model of running each test in a separate
      process it won't properly work anyways). Refactoring is done to
      reduce to the minimum associated code in sandstone.cpp and around
      main test loop.

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>