FW: Test mgmt refactor: control the quality from outside test set.
Test set only returns the next test it would run, agnostic of the
quality level requested and other factors that may influence the
decision on whether to run a test or not. One particular benefit this
offers is to provide meaningful diagnostics from the test run logic
(e.g. output the reason the test was skipped) from the test run logic as
opposed to test set container.

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>