tests/zlib: work around a race condition in x86_check_features()
This is a temporary fix to the race condition found in intel/zlib's
x86_check_features() function. The race condition is not present in
madler/zlib (official release) or in zlib-ng/zlib-ng. The issue is in
```c
    if (once != 0)
        return;
    once = 1;
```
because the `once` flag is set before the CPUID detection results are
stored, so some threads could being executing with uninitialised
(zeroed) state and then later see it changed. This can cause
```
'E> Failed at ../opendcdiag/tests/zlib/test.c:46: inflate failed: data error (-3)'
```
and possibly other unexpected results.

This is a work around and only works when linking statically to a build
of intel/zlib.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>