BATS: Use selftest_sigill instead of selftest_sigkill.
selftest_sigkill is not a valid test name on Windows. After the
refactoring opendcdiag.exe produces an empty output when an illegal test
name is specified. Before it was not touching the log file, but the test
was failing cause it was reading the output.yaml of the previous run
(note the loop over tests).

Also, make the output file unique.

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>